Compilation fixes (Win32)
git-svn-id: svn://svn.code.sf.net/p/glmixer/Source/trunk@694 a4deb95c-d308-4ec4-8dc1-6419bbeeb61a

+++ src/FFGLPluginSource.cpp
@@ -29,7 +29,7 @@ public:
         if (result!=NULL) {
             PluginInfoStructTag *plugininfo = (PluginInfoStructTag*) result;
             mapinfo.insert("Name", plugininfo->PluginName);
-            mapinfo.insert("Type", plugininfo->PluginType);
+            mapinfo.insert("Type", (uint) plugininfo->PluginType);
         }
         return mapinfo;
     }
